Safari-609.1.11.1/LayoutTests/svg/animations/svgenum-animation-13-expected.txt

SVG 1.1 dynamic animation tests

Test ChannelSelectorType enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS displacementMap.xChannelSelector.animVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.xChannelSelector.baseVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.xChannelSelector.animVal is SVGFEDisplacementMapElement.SVG_CHANNEL_R
PASS displacementMap.xChannelSelector.baseVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.xChannelSelector.animVal is SVGFEDisplacementMapElement.SVG_CHANNEL_R
PASS displacementMap.xChannelSelector.baseVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.xChannelSelector.animVal is SVGFEDisplacementMapElement.SVG_CHANNEL_G
PASS displacementMap.xChannelSelector.baseVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.xChannelSelector.animVal is SVGFEDisplacementMapElement.SVG_CHANNEL_G
PASS displacementMap.xChannelSelector.baseVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.xChannelSelector.animVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.xChannelSelector.baseVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.xChannelSelector.animVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.xChannelSelector.baseVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.xChannelSelector.animVal is SVGFEDisplacementMapElement.SVG_CHANNEL_A
PASS displacementMap.xChannelSelector.baseVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.xChannelSelector.animVal is SVGFEDisplacementMapElement.SVG_CHANNEL_A
PASS displacementMap.xChannelSelector.baseVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.xChannelSelector.animVal is SVGFEDisplacementMapElement.SVG_CHANNEL_A
PASS displacementMap.xChannelSelector.baseVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS successfullyParsed is true

TEST COMPLETE
